Refactor file edit logic to improve modularity and reusability 📦✨
