+++ packages/sample/genaisrc/python/genaiscript.d.ts
@@ -2197,6 +2197,8 @@ interface ChatGenerationContext extends ChatTurnGenerationContext {
         strings: TemplateStringsArray,
         ...args: any[]
     ): RunPromptResultPromiseWithOptions
+    defFileMerge(fn: FileMergeHandler): void
+    defOutputProcessor(fn: PromptOutputProcessorHandler): void
 }
 
 interface GenerationOutput {
@@ -2971,8 +2973,6 @@ interface ContainerHost extends ShellHost {
 interface PromptContext extends ChatGenerationContext {
     script(options: PromptArgs): void
     system(options: PromptSystemArgs): void
-    defFileMerge(fn: FileMergeHandler): void
-    defOutputProcessor(fn: PromptOutputProcessorHandler): void
     env: ExpansionVariables
     path: Path
     parsers: Parsers
